role based access

+++ SBS_Backend/sbs/src/main/java/com/securebanking/sbs/repository/UserRoleRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRoleRepo extends JpaRepository<UserRole, Long> {
+public interface UserRoleRepo extends JpaRepository<UserRole, Integer> {
 }
